replace buildV1 with buildV2, but still retain deprecated code
